requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.2
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
